Moved the showPulsatingDotAndPostHeader logic in the selectors and wrote tests for them
