+++ components/status_dropdown/index.jsx
@@ -13,8 +13,7 @@ import {openModal} from 'actions/views/modals';
 import {setStatusDropdown} from 'actions/views/status_dropdown';
 
 import StatusDropdown from 'components/status_dropdown/status_dropdown.jsx';
-import {getCustomStatus, isCustomStatusEnabled} from 'selectors/views/custom_status';
-import {showStatusDropdownPulsatingDot} from 'utils/custom_status';
+import {getCustomStatus, isCustomStatusEnabled, showStatusDropdownPulsatingDot} from 'selectors/views/custom_status';
 import {isStatusDropdownOpen} from 'selectors/views/status_dropdown';
 
 function mapStateToProps(state) {
